requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pydantic==2.7.1
openai==1.30.1
requests==2.31.0
python-dotenv==1.0.1
openenv-core>=0.2.1